requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
langchain-openai>=0.0.2
streamlit>=1.30.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
pydantic>=2.5.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.1.0
plotly>=5.18.0
docx2txt>=0.8
PyPDF2>=3.0.0
openlm>=0.1.0  # For OpenAI-compatible local LLM support

# Development dependencies
ruff>=0.1.0
pytest>=7.4.0
pytest-cov>=4.1.0
mypy>=1.5.0
types-requests>=2.31.0
types-PyPDF2>=3.0.0